sam3x8e: Move adc support to a new adc.c file
